config: update precache using main/4.21-RC response (#187)
Updates precache to output from a CloudStack repo main branch/4.21 env.
Some API/param descriptions now contain backticks, which break the format in precache.go.
Therefore, instead of adding listApis output as a string variable, store the output in a JSON file and use embed to get it in a Go variable.

---------

Signed-off-by: Abhishek Kumar <abhishek.mrt22@gmail.com>